From: rich@id.slip.bcm.tmc.edu.cdrom.com (Rich Murphey)
ld won't generate output if there are two .o files with the same
name.  It thinks they are multiply defined external references.  This
patch explictly allows multiple /file names/ with the same value.
This can happen if you do a two stage link or if you link in file
names that are identical to files in libc.

Otherwise ld.c exits with a status=1 and non error message.  Rich


git-svn-id: https://svn.freebsd.org/base/head@385 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f