>From: paul@u.tvt.se (Paul Pries) >Subject: Bad bug in kbdtables.h [FreeBSD]
I found a bug in /sys/i386/isa/kbdtables.h which contain the
different keyboard layouts for syscons.  This regards all tables exept
the Danish and US.  When compiling the kernel with any other keymapping
than Danish or none at all (US), you get an error that 'key_map' is undefined.
This is because there is a typo in the name of the struct containing
the tables, keymap intead of key_map.


git-svn-id: https://svn.freebsd.org/base/head@470 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f