>From David Greenman
Bruce Evans had limited the kernel virtual address space to not include the
last 4MB since it was not being used.  Other changes are being made that will
reloate the Alternate Page Directory Table (APDT) into this area so the limit
is being fixed to be the last virtual address. (Infact with this patch you
can now do that relocation)


git-svn-id: https://svn.freebsd.org/base/head@604 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f