Stop gap measure until we can get Bruce's driver debugged. --- From: sos@login.dkuug.dk (S|ren Schmidt) Subject: IDE-disk hangs - solution/patches NetBSD/FreeBSD Summary: fixes for lost interrupts with IDE disks Keywords: hanging-disk, IDE-disk, lost-interrupt
Due to "popular" demand I'm posting these patches to NetBSD/FreeBSD
instead of mailing them around the world :-)

As many have found out there is a problem when using IDE disks on
FreeBSD. Following is a patch that fixes the problem with lost intterrupts.
Both fixes is based on a patch posted here some month ago by
Stefan Behrens?? (sorry I've lost the original article). But anyway it
works (for me :-).
Basically it does a timeout on lost interrupts, starting the operation
again and logging and error message on the console.

It additionally makes the allready present while loop timeouts
independent of CPU speed, and adds minor numbers for easy access to
dos partitions.


git-svn-id: https://svn.freebsd.org/base/head@793 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f