New wd driver, based on Bruce Evans 'wx/altwd' driver which has passed enough tests to be considered more stable than current driver.
Lots of work by Bruce, David G., and Guido have gone into this version, and
more is to come in the future.

Support for multiple controllers is in, but doesn't work correctly with
different controllers (IDE AND MFM), but multiple alike controllers appears
to work.

Most of the stray interrupts problems should be fixed, although you will
get a couple 'extra interrupts' when disklabeling and on startup.


git-svn-id: https://svn.freebsd.org/base/head@933 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f