Fix spelling error in little-used target.
Build an intermediate object file even when there is only one source
file.  This costs a little space but saves time if the target is rebuilt
a lot, and it stops the target varying with the name of temporary
intermediates.

Use ${ECHO} instead of `echo' so that `make -s' is fairly quiet.

Use ${ECHODIR} instead of echo for printing directory names so that
`make -ss' is very quiet.


git-svn-id: https://svn.freebsd.org/base/head@2351 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f