From: hosokawa@mt.cs.keio.ac.jp (HOSOKAWA Tatsumi)
   One of the alpha testers (ETO, Toshihisa <eto@osl.fujitsu.co.jp>)
of my APM driver sent me a very small patch to if_ze.c for using IBM
PCMCIA Ethernet card II.  There are only a few difference between
Ethernet card I and II.  So we can use them both with this patch.  It
also includes a patch for PCIC of ThinkPad 230Cs (As long as I
remember, this model is available in Japan only. But it is very
popular subnote in Japan).
Submitted by:	hosokawa


git-svn-id: https://svn.freebsd.org/base/head@3759 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f