[Changes are now for BETA]
This attempts to make the installion a little less grating by not
clearing the screen every time a like-sized dialog is going to splat
on top of another one.  If you don't want the screen to clear, set
clear="", otherwise leave it set to (or set it back to) "--clear".

It can't be set to "" all the time since then small dialogs appearing
on top of big dialogs look funky.


git-svn-id: https://svn.freebsd.org/base/head@4326 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f