Added a new feature from Peter Eriksson's latest release of ypserv (0.13): register ourselves as an NIS version 1 UDP server to pacify older SunOS 4 ypbinds that seem to insist on having one around. All this does is allow ypserv to respond to DOMAIN_NONACK requests that are periodically transmitted by ypbind: the server will not actually work as an NIS v1 server in any other way.
Unlike the mainline code, which implements this as a compile-time
option, this feature can be turned on with the newly-added -k flag
at runtime.

Bunped version number to 0.13. (What the hell.)

Updated the man page to reflect this change, also made a couple of small
edits to reflect the recent changes in the /etc/rc* setup.


git-svn-id: https://svn.freebsd.org/base/head@7626 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f