Undo the stupidity I inflicted on these files and replace it with some (hopefully) less offensive stupidity:
If we detect that a user has loaded a module that fails to initialize
itself correctly, panic. There really isn't a safe way to recover from
something like this; we can't know that the module is bad until after
the entry point is called, by which time it's too late to do anything
about it.


git-svn-id: https://svn.freebsd.org/base/head@7947 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f