Set the return value (outb'd to the sequencer) for non zero status interrupts to zero to start with.  This is the value we return in most cases.  The sense code then outbs the apropriate value to cause sense retrieval.  The return value was uninitialized before this change (something that was an okay thing with the old sequencer code, but not with the semantics of the new).
This problem caused us to always retrieve sense upon recieving a non
zero status byte.  This is exactly what was happening for the "target
busy" status returned by exabyte tape drives when they rewind or
power on with a tape installed.  The request sense proved fatal.
This should fix the tape problems.


git-svn-id: https://svn.freebsd.org/base/head@8199 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f