Reviewed by:	sos Submitted by:	bde
rc.i386 failed messily when I used a serial console.

Editors note:
Use file redirection so that all the kbdcontrol and vidcontrol
commands act upon /dev/ttyv0 instead of stdin.  Though this feature
is not documented it is the intended behavior of {vid,kbd}control
and shall be documented in the future as such.


git-svn-id: https://svn.freebsd.org/base/head@8767 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f