Don't trash the denode cache entry for the source when renaming a directory in the newparent case.  For directories, the pseudo-inode number (de_dirclust, de_diroffset) is canonicalized to be (first cluster in directory, 0), so it doesn't change when directories are renamed.  This was only partly handled by not doing anything for directories in reinsert().
It is hard to demonstrate bad effects of this bug.  You can see an
unnecessary cache miss by tracing msdosfs_hashget().  The cache
entry is usually left on the wrong hash chain so it is very
unlikely to match anything, and bogus cache entries go away in
msdosfs_reclaim().

Submitted by:	Bruce Evans


git-svn-id: https://svn.freebsd.org/base/releng/2.0.5@8972 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f