Patch to fix PR #518. In a system with no NIS, passwd will complain that it can't contact an NIS server when asked to change the password of an invalid user. It should say 'unknown user' instead.
The fix is to check for the _PW_KEYYPENABLED flag in the password
database and only roll over into the yppasswd stuff if the flag
is enabled (this means passwd will not behave as yppasswd if
there are no +::::::::: entries in /etc/master.passwd). If
NIS is enabled but the user says 'passwd -l foouser' where
foouser exists in the NIS maps, but not in /etc/master.passwd,
we also say 'unknown user.' This is so we don't outsmart ourselves:
specifying the -l flag restricts passwd to the local password database
even if NIS is enabled.

This change should probably be merged into 2.1.


git-svn-id: https://svn.freebsd.org/base/head@9239 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f