Reduce timeout frequency from `hz' to 0 if no ports are open or to 1 if no ports are active, provided there are no polled ports and no `LOSESOUTINTS' ports.  Do a little more in the interrupt handler instead. This is a little less efficient if there are are many active ports but a little more efficient otherwise.  Polled ports are ones with no irq specified (as before).  `LOSESOUTINTS' ports are ones with 0x08 set in their config flags.  Unless this flag is set, it will now take up to one second to recover from lost output interrupts, if any.  Some 8250s and 16450s lose output interrupts.
Improve output buffering: copy the clist buffer to 2 linear buffers if
necessary and possible instead of to 1.  Handle an arbitrary queue of
buffers in the interrupt handler.  Check for waking up sleepers after
copying characters out of the clist buffer instead of before.

Delay translation of TIOCM_DTR to MCR_DTR etc. so that the top level
routines are more machine independent.

Fix bogus device register in unused code.


git-svn-id: https://svn.freebsd.org/base/head@9307 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f