Total rewrite of the dataphase sections of the sequencer.  This was done to replace the very poor, original implementation of Scatter/Gather operations.
Use a bit (that was freed up with the rewrite above) in the SCB control byte
to designate commands that should allow disconnection.  The kernel driver
makes this decision now instead of the sequencer since the sequencer can't
do the indexing very efficiently.

This commit drops the sequencer from 426 instructions to 390 most likely
freeing enough space to do a target mode implementation.


git-svn-id: https://svn.freebsd.org/base/head@9928 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f