Fixup the "ld.so failed" message for the case when ld.so finds undefined symbols.
An easy example to see this is to develop an X program which links
against Xt, but doesn't add -lX11 to the link line.  It will link fine,
but cause run-time errors by ld.so because of missing symbols used by Xt
defined in X11.  This patch makes the errors more readable.

Submitted by:	jdp@polstra.com (John Polstra)


git-svn-id: https://svn.freebsd.org/base/head@11039 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f