I have applied my last changes correcting the bug in dgbselect() to it and here is the patch.  Submit it please. Thank you!
BTW, there is a new option "NDGBPORTS". By default it is equal to
NDGB*16 and means the number of ports of all Digiboards for which the
tty structures are reserved. It can be set to the real value in config-file
like:

       options "NDGBPORTS=8"
Submitted by:	Serge A. Babkin <babkin@hq.icb.chel.su>


git-svn-id: https://svn.freebsd.org/base/head@11188 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f