Add a few more heuristics to modload: 1: generate the outfile in /tmp if it's not specified explicitly. 2: if the outfile was implicitly placed in /tmp, automatically remove it.
This means that you can type: modload /lkm/ipfw_mod.o and it'll work, it
wont try and write to /lkm, and it wont leave the (normally) useless
symbol file.

This should not interfere with things like ibcs2 and atapi, which
export some symbols from one LKM to the other by leaving the symbol file.


git-svn-id: https://svn.freebsd.org/base/head@11879 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f