Oops, forgot the following log message in the previous commit:
Included <sys/sysproto.h> to get central declarations for syscall args
structs and prototypes for syscalls.

Ifdefed duplicated decentralized declarations of args structs.  It's
convenient to have this visible but they are hard to maintain.  Some
are already different from the central declarations.  4.4lite2 puts
them in comments in the function headers but I wanted to avoid the
large changes for that.


git-svn-id: https://svn.freebsd.org/base/head@12223 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f