This driver supports the Arnet SYNC/570i ISA cards that is based on the HD64570 chip. Both the 2 and 4 port cards is supported and auto detected.
Line speeds of up to 2Mbps is possible. At this speed about 85% of the
bandwidth is usable with 486DX processors.

The standard FreeBSD sppp code is used for the link level layer. The
default protocol used is PPP. The Cisco HDLC protocol can be used by
adding "link2" to the ifconfig line in /etc/sysconfig or where ever
ifconfig is run.

At the moment only the V.35 and X.21 interfaces is supported. The others
may need tweaks to the clock selection code.

Submitted by: John Hay <jhay@mikom.csir.co.za>


git-svn-id: https://svn.freebsd.org/base/head@12437 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f