Get rid of two and a half printf in the kernel. Add more features to the one remaining to handle the job: 	+	signed quantity. 	#	alternate format 	-	left padding 	*	read width as next arg. 	n	numeric in (argument specified) default radix.
Fix the DDB debugger to use these.
Use vprintf in debug routine in pcvt.

The warnings from gcc may become more wrong and  intolerable because
of this.

Warning:  I have not checked the entire source for unsupported or
changed constructs, but generally belive that there are only a few.

Suggested by: bde


git-svn-id: https://svn.freebsd.org/base/head@13446 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f