Some news for syscons (long overdue):
Real support for a Textmode mousecursor, works by reprogramming the
charset. Together with this support for cut&paste in text mode.
To use it a userland daemon is needed (moused), which provides
the interface to the various mice protokols.
Bug fixes here and there, all known PR's closed by this update.


git-svn-id: https://svn.freebsd.org/base/head@16564 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f