Add a facility for a ``slow'' CTM delta queue, allowing `x' number of delta's to be mailed out every hour (or however often you schedule the cron job).
ctm_dequeue is the cron job which takes the stuff from the
queue directory and punts it into sendmail. The chunks of
the deltas (and the complete deltas if they are that small)
are sorted into order before being dispatched, so the people
subscribing should still get the bits in the right order.

The changes to ctm_smail should be fairly safe as they won't be
activated unless you go for the new queue directory option.


git-svn-id: https://svn.freebsd.org/base/head@16880 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f