Repository copy libfakegnumalloc to libgnumalloc and remove the symlink stuff and detection for the "gnumalloc" port which doesn't exist and nobody has cared enough about for the past 6 months or so to implement.
As has been pointed out to me (quite a few times) in email, the people
that had been bitten by the changes had failed to follow the instructions
about updating /etc/rc.

Bruce also pointed out that after my last commit, it was no longer
removing /usr/lib/libgnumalloc.so.2.0 as it should have been.

Hopefully this (and the comments in the Makefile) should defuse the problem
a bit.


git-svn-id: https://svn.freebsd.org/base/head@18506 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f