Don't allow empty pathnames.  POSIX standard.
Most of the standard utilities that depended on (or were broken in
a different way by) the old behaviour of interpreting "" as "."
were fixed a year or two ago.  There is still a fairly harmless
bug in tar and a harmless bug in gzip.  Tar apparently replaces
"/" by "" when it strips leading slashes.


git-svn-id: https://svn.freebsd.org/base/head@20069 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f