Suggested by: msmith Add a little more sophistication to the md5 grep command.
Change the md5 checksum logic a bit.  Now, the message is printed
out for every successfully/unsuccessfully matched checksum, and it
aborts at the end if there was a mismatch.  Also, make missing checksum
and IGNORE file inconsistency fatal, as there is now no reason to have
a missing checksum.


git-svn-id: https://svn.freebsd.org/base/head@20843 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f