Documentation update.
Remove a duplicate entry for ahc0 in hardware.hlp (closes PR #
docs/2282), a few additions all over the place while i was at it.

Tried to add as much of the important highlights for the new features
in 2.2 into the release notes.  I still think that many things are
missing here, so folks, if you imported something great that's in 2.2
but not in 2.1.X, please review this section and send us your
additions!

Submitted by:	seki@sysrap.cs.fujitsu.co.jp (for the ahc0 dup)


git-svn-id: https://svn.freebsd.org/base/head@20928 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f