Provide weak references to functions that will replace certain names in the 4.9.4-P1 resolver that we have in libc as of 2.2 and current 3.0.
When 4.9.5-P1 is imported, certain well known functions will "go away"
and be replaced by functions under new names.. which just happen to be
what these weak references point to.

The result should be that while 2.2 and 3.0 keep using libc.so.3.0, they
should be binary compatable even though they have a different version of
the resolver in their libc's.  Of course, if the new resolver makes it into
2.2-RELEASE (which it probably should if it's stable), then this is
irrelevant. :-)


git-svn-id: https://svn.freebsd.org/base/stable/2.2@21054 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f