- major overhaul to make this deal with unlimited fd's. - kill non-FD_SETSIZE code
Obtained from: a diff of FreeBSD vs. OpenBSD/NetBSD rpc code.

Note, there was a nasty bug with our old code here.  It would trash the
stack if a fd > 31 was passed in.  It was using a "long" as though it
was an "fd_set", ie: it was assuming that a long was 256 bits wide. :-(
This has been lurking here for a while, since the FD_SETSIZE #ifdef's
were first implemented.


git-svn-id: https://svn.freebsd.org/base/head@21087 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f