Add OpenBSD support.  Change some comments to reflect the fact.  Also:
(1) Change commented out MAINTAINER to FreeBSD_MAINTAINER and
    OpenBSD_MAINTAINER.  These are not comments anymore, so we may
    even use it in the future.

(2) Instead of the ".if ${OPSYS} = "NetBSD" hack, use ".if exists()"
    to find the location of md5 an tar.  Play similar trick for fetch
    (OpenBSD uses /usr/bin/ftp which groks http: addresses).

This commit includes most of the changes made in 1.242 (although many
of them are done differently after more discussion).  One thing that
is conspicuously missing is NOMANCOMPRESS, which has been postponed
until Warner figures out what exactly the situation is on the OpenBSD
ports paradigm.  (In a nutshell, we can't just define NOMANCOMPRESS in
this file even if uncompressed manpages is the default for OpenBSD,
because that will take away the ability of individual users to select
manpage compression.)

Reviewed by:	imp@openbsd.org


git-svn-id: https://svn.freebsd.org/base/head@21604 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f