Fix an exploitable buffer overflow condition.  From l0pht:
Modstat is sgid kmem which is really handy to become if you feel like
looking through /dev/mem and /dev/kmem (gee, wonder what you might
want to do that for ). Like just about everything else under the sun
it has a buffer overflow problem. The problem exists in the dostat()
routine where an arbitrary sized string is shoved into sbuf.name
through a strcpy().

<http://www.l0pht.com/advisories.html>

This change was committed to 2.2 as rev. 1.4 of modstat.c by joerg, and is
being duplicated verbatim in 2.1-STABLE -JG


git-svn-id: https://svn.freebsd.org/base/stable/2.1@22311 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f