23:59's mega-merge for ppp(8).
I've merged the three changes:

    re-open on HUP (+doc)
    Working -background option
    pending sig changes

and enclosed patches to get from the current RELENG_2_2.
The pending signal stuff is the most important, without it it's possible
to corrupt the caused variable, resulting in a call to (*NULL)() :(

(I'm not totally comfortable with the SIGHUP changes, but they are not
very harmful either. -- Joerg)

Submitted by:	brian


git-svn-id: https://svn.freebsd.org/base/stable/2.2@23896 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f