Fix the gdb executable modify problem.  Thanks to the detective work by Alan Cox <alc@cs.rice.edu>, and his description of the problem.
The bug was primarily in procfs_mem, but the mistake likely happened
due to the lack of vm system support for the operation.  I added
better support for selective marking of page dirty flags so that
vm_map_pageable(wiring) will not cause this problem again.

The code in procfs_mem is now less bogus (but maybe still a little
so.)


git-svn-id: https://svn.freebsd.org/base/head@24666 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f