Generate mount*.[ch] and nfs_prot*.[ch] from the .x files.
Delete bogus local versions of mount*.[ch] and nfs_prot*.[ch].

Use v3 protocol by default for NFS mounts.  If v3 is not supported, v2 is
used automatically.  Add a new mount options for NFS, 'nfsv2' to allow the
administrator to force the use of the older protocol.

Document the new mount options and fix a couple of markup problems.


git-svn-id: https://svn.freebsd.org/base/head@25001 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f