YAMFC:  I did the later commit which Bruce mentioned in the mailing list, but I noticed this one wasn't brought in, and was specifically mentioned in the logs.  I've tested this and the previous vix on my box and it appears to work as well as before. :)
revision 1.42
date: 1996/11/13 18:31:57;  author: bde;  state: Exp;  lines: +28 -17
Don't fiddle with RTS if RTS flow control is off.  This gives applications
almost complete control over RTS (control of its initial value is still
missing).

This fixes PR 1644 for sio.

The author of PR 1644 wants it in 2.1.6 and 2.2.  This may be safe since
the complications are only in rarely used cases that I hope I've covered.


git-svn-id: https://svn.freebsd.org/base/stable/2.1@27485 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f