Call poll(2) from within the resolver but adapt to older kernels without it if necessary.  This removes the need to malloc large fd_set's for selecting on high fd's (larger than FD_SETSIZE at libc compile time).
The syscall adaptive stuff only happens on the very first call.  SIGSYS
is masked, and if the call to poll fails with ENOSYS, then we use select
for the life of the program.  If poll does not fail with ENOSYS, then we
always use poll and skip the once-off signal masking gunk.

This may be overkill, but it saved my neck a few times while working on
multiple different sets of kernel sources, some with poll, some without.


git-svn-id: https://svn.freebsd.org/base/head@29380 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f