In the true spirit of 'code freeze^H^H^H^H^H^Hslush', bring over the updated de driver from -current, minus the include bashing, the multicast interface changes and the pci interface changes.  This essentially leaves us with Matt's 970703 driver but with the tx descriptor page wrap bug fixed.
However, there are some known problems remaining.  If the new problems are
found worse than the old ones, and they are not found and fixed prior to
release, this will be reverted to the old driver.

The (known) problems are:
  - autoselect is a bit flakey on older cards with BNC in use.  An explicit
    setting of the media modes is thought to work.
  - There were some reports of OACTIVE getting stuck "on" and/or
    "Transmitter underflow" requiring an ifconfig down and up to unjam it.
    More details are needed.

The old driver simply doesn't work on a good number of newer cards.  We
can't afford to not give this a try.

BETA-"no major changes"-rule-violation OK'ed by: jkh
Approved-with-grumbles-by: dg


git-svn-id: https://svn.freebsd.org/base/stable/2.2@30254 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f