Handle machine-dependent (stdlib) sources more automatically.
This fixes bugs in the manual handling.  abs.[cS] was handled too
specially and the wrong (.c) variant for each of div.[cS], labs.[cS]
and ldiv.[cS] was added to SRCS.  This caused the .c variant to be
used if `depend' was made and the .S version to be used otherwise.


git-svn-id: https://svn.freebsd.org/base/head@30484 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f