Make emacs work again.  This is a workaround for the fact that the emacs a.out file, self-generated by emacs's "unexec" function in "unexsunos4.c", is invalid.  In particular, its "_end" symbol has the wrong value.  The dynamic linker was using the value of that symbol to initialize its sbrk break level.
The workaround is to peek at the executable's a.out header in
memory, and calculate what "_end" should be based on the segment
sizes.

I will work out a fix for emacs and send it to the FSF.  This
dynamic linker workaround is still worthwhile, if only to avoid
forcing all emacs users to build a new version.

Note: xemacs gives a bogus warning at startup, for related reasons.
The warning is harmless and can safely be ignored.  I will send a
patch to the xemacs maintainers to get rid of it, and meanwhile
add a patch file to our port.


git-svn-id: https://svn.freebsd.org/base/head@31546 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f