Fixed DPADD.  ${LIBGCC} must be added explicitly since the default linkage rule is overridden.  The -L option must be in ${DPADD} so that `make checkdpadd' works.  Actually use ${DPADD}.
FIxed missing dependencies for doscmd.  Use ${LIBCRT0} instead of a
hard-coded path in the rule for doscmd.

Added comments about the kludges used to build 2 binaries and 2 data
files in one directory.  It shouldn't be done this way.  The dependencies
on sources took extra work to get right, and the dependencies on objects
are still broken (one set is missing and the other has the wrong libs).

Fixed some style bugs while I'm here:
- don't override the (correct) default for MAN1.
- use `beforeinstall', not `afterinstall' to install auxiliary files.
  `afterinstall' is for fixing messes made by `install'.


git-svn-id: https://svn.freebsd.org/base/head@31776 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f