Move the trampolines for dlopen and related functions from crt0.o into libc.  This reduces the size of every dynamically linked executable by 248 bytes, and it reduces the size of static executables by a lesser amount.  It also eliminates some global namespace pollution.
With this change in place, the source for dlfcn.h should probably
be moved to "/usr/src/include".  I'll save that for another day.

Compatibility note:  Programs which use dlopen, if compiled on
systems with this change, will not run on systems with a libc from
prior to this change.  Very few programs use dlopen, so I think
that is OK.


git-svn-id: https://svn.freebsd.org/base/head@33180 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f