Update timer0_prescaler_count before calling hardclock() while timer0 is "acquired".  This fixes a TSC biasing error of about 10 msec when pcaudio is active.
Update `time' before calling hardclock() when timer0 is being released.
This is not known to be important.

Added some delays in writertc().  Efficiency is not critical here, unlike
in rtcin(), and we already use conservative delays there.

Don't touch the hardware when machdep.i8254_freq is being changed but
the maximum count wouldn't change.  This fixes jitter of up to 10 msec
for most small adjustments to machdep.i8254_freq.  When the maximum
count needs to change, the hardware should be adjusted more carefully.


git-svn-id: https://svn.freebsd.org/base/head@33309 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f