Add join to the list of bootstrap tools and remove the cpp script because we set CPP explicitly.
Ensure that the make flags are used for depends too.

After building the build tools with LMAKE (which allows installed tools
to be used), rebuild the same tools with XMAKE (which does not use
the installed tools, headers or libraries) to check that the tools can
build themselves and that there are no missing tools.


git-svn-id: https://svn.freebsd.org/base/head@35901 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f