The aout->elf transition build from a 2.2.5/6/7 system requires that a kernel be built and installed at just the right time. If the system were to be rebooted after installing the 3.0 binaries without the new kernel, the system is hosed.
The addition of this config file reserves the name of the config file
that the upgrade procedure uses. If an upgrade is to be performed on
a 2.2.5/6/7 system system for which a GENERIC kernel is inadequate,
this config file must be replace with a suitable one before performing
the upgrade.

On systems which report 'uname -r' as 3.0-CURRENT, the kernel build
is not performed and this file is ignored.


git-svn-id: https://svn.freebsd.org/base/head@38636 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f