[As I was saying]
I believe this by bde from bsd.prog.mk still applies:

To get the fixes and simplifications, application Makefiles should put the raw
lex and yacc source names (foo.l and bar.y) in SRCS and not put the names of
any generated files in SRCS or CLEANFILES.

If you want to regress to generating foo.c from foo.y via y.tab.c for crufty
applications that want a y.tab.h file.  This want must be specified by putting
y.tab.h in SRCS (and defaulting to or putting -d in YFLAGS).  This only works
if there is only one yacc parser, of course.

Complications: if bar.y is in SRCS and -d is in YFLAGS, then bar.h will be
generated whether or not it is used, even if this clobbers a real source file.


git-svn-id: https://svn.freebsd.org/base/stable/2.2@38735 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f