Fix the MS-CHAP support.  The code was passing a bit count to MD4Update(), but our version in libmd expects a byte count.
This code is not currently compiled or linked into pppd, so I'm
reasonably sure I didn't break anything. :-)  I added the necessary
statements to the Makefile, but left them commented out because we
are in feature freeze.  When the code is enabled, we must be careful
to build it only if the DES library is available.


git-svn-id: https://svn.freebsd.org/base/head@40236 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f