Update the cd(4) man page for the CAM cd driver.
This includes a description of the changer timeout kernel options and
sysctl variables.  I didn't check to make sure the ioctl descriptions are
up to date; that will come sometime later.  (The ioctls haven't changed in
the CAM driver, but I'm not sure if the man page was in sync with even the
old driver.)


git-svn-id: https://svn.freebsd.org/base/head@40246 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f