Implement a simple LRU block cache.  By default this is initialised to 16k, and will bypass transfers for more than 8k.  Blocks are invalidated after 2 seconds, so removable media should not confuse the cache.
The 8k threshold is a compromise; all UFS transfers performed by
libstand are 8k or less, so large file reads thrash the cache.
However many filesystem metadata operations are also performed using
8k blocks, so using a lower threshold gives poor performance.

Those of you with an eye for cache algorithms are welcome to tell me
how badly this one sucks; you can start with the 'bcachestats' command
which will print the contents of the cache and access statistics.


git-svn-id: https://svn.freebsd.org/base/head@40834 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f