Straigthen out the use of the tls and tlf callbacks.
Not tested on the if_sr, if_cx and if_ar drivers, but
expected to work just the same as it used to.

Any users of these drivers (or even better: donors
of hardware for them) please contact phk@freebsd.org
so we can test the next batch of changes to if_sppp.


git-svn-id: https://svn.freebsd.org/base/head@41881 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f