Declare size_t and ssize_t if they are not already declared, so that <kvm.h> is self-sufficient again.
Moved typedefs and forward struct declarations out of __BEGIN_DECLS/
__END_DECLS.

Don't comment out the prototype for kvm_uread().  This was a 4 year
old kludge for previous breakage of self-sufficiency.  The prototypwe
was broken instead.

Fixed bitrot (const poisoning) in the type of kvm_uread().

Fixed order of the declaration of kvm_uread().


git-svn-id: https://svn.freebsd.org/base/head@41882 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f