The Conner CFP2107 is a fixed-media drive, not removable media.  This fixes the quirk that disables tagged queueing for those drives.
Also, silence a warning by disabling xpt_for_all_targets() and
xpt_for_all_periphs().  These two functions are not currently used, but
they should not be removed.  They're part of a set of functions that
provide a way to execute a function for every {bus,target,device,periph} in
the system.

If anyone needs to use either function in the future, they can be
un-#ifdefed.


git-svn-id: https://svn.freebsd.org/base/head@42319 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f