A slight bit of code and doco cleanup, but mostly:
Add the much-needed and long-desired ability to force the log rotation
to take place on a specific day or time, and document same.  This works
by extending the syntax of the `when' field to allow a restricted
ISO 8601 time specification, and performing the log rotation if newsyslog
is run during the same hour as is specified.


git-svn-id: https://svn.freebsd.org/base/head@43071 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f