Removed all traces of `p_switchtime'.  The relevant timestamp is per-cpu, not per-process.  Keep it in `switchtime' consistently.
It is now clear that the timestamp is always valid in fork_trampoline()
except when the child is running on a previously idle cpu, which
can only happen if there are multiple cpus, so don't check or set
the timestamp in fork_trampoline except in the (i386) SMP case.
Just remove the alpha code for setting it unconditionally, since
there is no SMP case for alpha and the code had rotted.

Parts reviewed by:	dfr, phk


git-svn-id: https://svn.freebsd.org/base/head@44327 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f