On FreeBSD/alpha, ps(1) does not correctly report process start times and CPU runtime because it can't access the user area via /proc/<pid>/mem. This is because the uarea is not mapped into the process address space at USRSTACK on the alpha like it is on the x86.
Since I'm haven't been able to wrap my brain around the VM system enough
to be able to figure out how to achieve this mapping, and since it's
questionable that such an architectural change is correct, I implemented
a workaround to allow ps(1) to read the uarea from /dev/kmem using
kvm_read() instead of from the process address space via kvm_uread().
The kludge is hidden inside #ifdef __alpha__/#endif so as not to impact
the x86. (Note that top(1) probably uses this same gimmick since it works
on FreeBSD/alpha.)

Reviewed by: dfr


git-svn-id: https://svn.freebsd.org/base/head@45227 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f