Build the functionality of the wdc_p hack into the ide_pci.c code. All it did was match a specific device ID and turn on a quirk for the wdc driver.
Incidently, at line 1462 there is a return that prevents the generic
ide_pci code from trying to look at the device.  I'd be interested
to know if we can take out the return and let the generic code "see" it.
I've left the return in because that's the way it worked before.

(Be sure to rerun config after cvsup or you'll get undefined files!)


git-svn-id: https://svn.freebsd.org/base/head@45667 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f