TARGET was being defined wrong, causing ld to always fail.  It is the name of the bfd target, not the gnu-standard target name.  Corrected to be elf32-{big,little}mips from mipse[bl]-unknown-freebsd.
DEFAULT_EMULATION was bogusly defined, causing ld to always fail (this
was masked by the TARGET bogosity).  Define correctly as elf32bmip and
elf32lmip.  Mips doesn't follow the same conventions as i386 and alpha
do in this area.

ld now appears to work correctly for the uncommitted mips changes to
egcs.


git-svn-id: https://svn.freebsd.org/base/head@45904 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f