Make DEVFS use PHK's specinfo struct as the source of dev_t and devsw.
In lookup() however it's the other way around as we need to supply the
dev_t for the vnode, so devfs still has a copy of it stashed away.

Sourcing it from the vnode in the vnops however is useful as it makes
a lot of the code almost the same as that in specfs.


git-svn-id: https://svn.freebsd.org/base/head@50334 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f