MFC: [ from Justin's email: "Almost everyone using the bus-dma interface copied a bug from my code. I fixed this in -current a month ago but forgot to MFC them.  I don't know how the old code ever worked... 8-)
There is also two fixes for detaching devices in here.  One is for
disconnecting a device with bus-dma resources (busdma_machdep.c) and
the other is for cam queue resources.  These two fixes have been in
-current for over a month (Nick Himba found them using USB devices).

And finally, there is a fix for a compile warning in the advansys
driver.

I may not be around for much of today, so feel free to commit these
if they are approved." ]

Submitted by:   gibbs


git-svn-id: https://svn.freebsd.org/base/stable/3@51256 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f