Slight reordering of /etc/rc order: It now goes
...
recover vi
msgs/bounds
update MOTD
rc.{arch}
rc.devfs
rc.local
$local.startup
securelevel

The motive behind this is to delay rc.{arch} as long as possible,
as it loads the screensaver, which kills the splash screen.
But at the same time, it can't be done after rc.devfs, as
that might depend on arch specific actions. rc.local and local.startup
also should come after these, and securelevel must be last of all.


git-svn-id: https://svn.freebsd.org/base/head@51617 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f