Revert to rev 1.91.  Revs 1.9{2,3} were causing problems for people with pre-Aug 4.0-CURRENT worlds and those with pre-GCC 2.95.2 worlds.
The problem with pre-Aug worlds is the installed Byacc and Bison doesn't
have necessary changes to compile either GCC 2.95 or EGCS 1.1.x.

The problem with pre-GCC 2.95 worlds is libgcc is built with the wrong
compiler.  See rev 1.17 of src/gnu/lib/libgcc/Makefile (which used to live
in src/gnu/usr.bin/cc/libgcc) + commit messge for details of the requirements.


git-svn-id: https://svn.freebsd.org/base/head@53653 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f