Revert previous commit, and
o  Add genassym to the list of cross-tools
o  Remove sh hashing work-around, we don't need it anymore
o  Clean more directories in WORLDTMP when NOCLEAN is specified

The sh hashing work-around is not needed anymore, because we don't
trigger the bug anymore.

When NOCLEAN is not defined, we wipe out the complete WORLDTMP,
including the object directories of the tools we have built. When
NOCLEAN is defined, we remove anything that we install anyway, which
is usr/bin, usr/games, usr/include, usr/lib and usr/sbin.


git-svn-id: https://svn.freebsd.org/base/head@55025 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f