truss /usr/bin/su 	login (or not if root) 	then exit the shell
truss will get stuct in tsleep

I dont know if this is correct, but it fixes the problem and
according to the commends in pioctl.h, PF_ISUGID is set when we
want to ignore UID changes.

The code is checking for when PF_ISUGID is not set and since it
never is set, we always ignore UID changes.

Submitted by:	Paul Saab <ps@yahoo-inc.com>


git-svn-id: https://svn.freebsd.org/base/head@55338 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f