Minor enhancement: set the 'dual link LED' bit in the auxmode register of the Broadcom BCM5201 PHY on the LinkSys USB100TX adapter so that the link LED correctly (lights up amber for 10mbps link, green for 100mbps link).
Note that the sticker on the bottom of the adapter says amber for 10
and green for 100, but the appendix in the manual that comes with
the adapter says green for 10 and amber for 100. Given that there doesn't
seem to be any way to make the hardware produce the latter combination,
I think it's safe to say the sticker is right and the manual is wrong.
I'm just shocked, shocked I tell you.


git-svn-id: https://svn.freebsd.org/base/head@56052 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f