Fix a bug that could crash the system if you press ^T while a slower system is slowed down and in the right spot (a race condition in fork()).
The "previous time" fields have moved from pstat to proc.  Anything which
uses KVM needs to be recompiled with a new libkvm/headers.

A couple wacky u_quad_t's in struct proc are now u_int64_t (the same, but
according to lack of 'quad's in proc.h and usage in kern_resource.c).
This will have no effect on code.

This has been make-world-and-installed-new-kernel-which-works-fine-tested.

Reviewed by:	bde (previous version)


git-svn-id: https://svn.freebsd.org/base/head@56766 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f