Approved by: jkh Reviewed by: joerg
The isdnd is able to listen on a socket for isdnmonitor to connect to
it to remotely control it (similar to ppp and pppctl). When this is
enabled in the isdnd config file, it will fail currently because isdnd
is started before the network interfaces are configured.
It is necessary to move the isdnd start after the ifconfig of the network
interfaces, then this problem will not occur.


git-svn-id: https://svn.freebsd.org/base/head@57012 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f