Snap shot check in.
The code doesn't compile (I'll do that in a minute), but functionally
it gets to receveing the interrupt from the start or join net command.

There is a lot of code to deal with the limitations of the PCCard memory
mapping code - I'm changing some of it hence the check in to preserve something
that worked.


git-svn-id: https://svn.freebsd.org/base/head@57515 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f