Fix crashes on card eject for pccard modems.  We check for NULL when we get the com address.  If so, we go ahead and return.  Bruce thinks there's a bug in the pccard layer that it terminates devices with extreme prejustice rather than letting them deside for themselves when to terminate (and he's likely right).  This fix doesn't change that, but instead works around it by checking for NULL pointers at more places than before.
The detach routine still calls functions at interrupt level that
aren't reentrant.  In theory this could cause a problem, but none
showed up in practice.  Future versions should correct this problem,
likely by making the detach process a thread/process at the pccard
level.  NEWCARD will do this, and the current pccard layer should
likely be modified to that as well, should it live long enough.

A few style nits of the same form that were in my original patch sent
off to bde were also fixed as part of this process.  Mostly use of
!ptr and return ENOPARENS.

This should prevent a crash on suspend with an active ppp link as
well, but that wasn't tested.

Reviewed by:	bde
Approved by:	jkh


git-svn-id: https://svn.freebsd.org/base/head@57915 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f