Don't forget to clean "grog".
Don't use MANDEPEND.  It hasn't had anything to do with dependencies for
5-6 years, but is still being used, mainly in groff/*/Makefile, where it
amounts to just a macro giving the list of generated man pages.  Since
all man pages in groff are generated (from .man to .[1-9]), it's simpler
to use the source names ({$MANX}) to give the list.

Fixed some other style bugs.


git-svn-id: https://svn.freebsd.org/base/head@60600 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f