Handle watchdog timeouts better. We can't really call the foo_init() routines from foo_watchdog() because foo_watchdog() is called at interrupt context, and that's a no-no due to the way the USB stack is currently set up.
What we do now is call the TX end of frame handler manually to clear
the completed transmission, then check the send queue and send off
any frames that are pending.

Also turned off the interrupt pipe stuff in if_aue, since it appears
to tickle a bug in the USB stack that I haven't found yet.


git-svn-id: https://svn.freebsd.org/base/head@61155 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f