If the kevent() registration fails (probably due to the underlying filesystem not being kq-aware), then fall back to using sleep.  This allows tail to work with NFS filesystems again without chewing up CPU time.
When given the -F flag, resort to sleep/stat after the file was moved
or deleted.  This allows a window where the file being tailed does not
exist at all, which is typically the case during log rotation.  Switch
back to using kq (if possible) after the file is reopened.


git-svn-id: https://svn.freebsd.org/base/head@61964 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f