MFC: camcontrol.c rev 1.25, camcontrol.8 rev 1.23
Eliminate the default device name (da) and unit number (0) in camcontrol.

This may break some scripts, but with the number of ways users can damage
a system with this tool, it's important to make sure they specify which
device they want to talk to.

Suggested by:	joerg
Tested for RELENG_4 by:	markm, nectar, will


git-svn-id: https://svn.freebsd.org/base/stable/4@62371 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f