MFC: Block signals while holding the exclusive lock to avoid locking against self.  The revisions merged are:
    alpha/lockdflt.c:	1.5 -> 1.6
    i386/lockdflt.c:	1.5 -> 1.6

I discussed this merge with the release engineer and we agreed that
this particular merge would best be done as soon as possible, rather
than waiting until just before the 4.1 code freeze when it couldn't
receive the benefit of testing with the ports it is intended to
fix.  (It should fix Mozilla and other user threads based programs.
And it should fail to break any linuxthreads based programs.)

Approved by:	jkh


git-svn-id: https://svn.freebsd.org/base/stable/4@63370 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f