A quick fix to get around  a problem (described below) with cia based machines.  The patch uses an existing global variable in place of the newbus accessor to get at use_bwx.
This is a quick fix to get miatas booting again; somebody
with more newbus skills than I can muster will have to correct it.

Matt Jacob's description of the problem from the -alpha list:

The IVAR accessor stuff for pcib is incompletely specified for CIA. There's
only one accessor defined, and that's to get the BUS instance number.
<..>
The device methods that try and get at the use_bwx get overriden because
there's only one ivar for CIA's pcib, and that's for hose #, and it's always
zero.


git-svn-id: https://svn.freebsd.org/base/head@65352 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f