MFC:
Remove hash table lookups and maintenance from chgproccnt() and chgsbsize()
and chase pointers stored in pcred and ucred instead for better performance
and to avoid these operations in interrupt context which could possibly
cause panics.

Because the pcred and ucred structures changed size, libkvm and friends
will need to be rebuilt.


git-svn-id: https://svn.freebsd.org/base/stable/4@65586 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f