Add constructors to crude structure support. Rework some of the code into a more modular interface, with hidden vocabularies and such. Remove the need to a lot of ugly initialization.
Also, add a few structure definitions, from stuff used on the C
part of loader. Some of this will disappear, and the crude structure
support will most likely be replaced by full-blown OOP support
already present on FICL, but not installed by default. But it was
getting increasingly inconvenient to keep this separate on my tree,
and I already lost lots of work once because of the hurdles, so
commit this.

Anyway, it makes support.4th more structured, and I'm not proceeding
with the work on it any time soon, unfortunately.


git-svn-id: https://svn.freebsd.org/base/head@65615 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f