Shut up some debugging messages that appear to have been left on by default.
Fix the identify/probe/attach cycle somewhat.  Now in acpi_identify we
parse enough of the RSDT/FACP to find the communications port and
interrupt so that we can establish these as resources on the acpi device.

Pass the softc to acpi_queue_event so that it doesn't dereference NULL
when acpi_intr is invoked in the attach routine (because devclass_get_softc
isn't going to work yet).

Account for our allocated resources properly so that we can free them
if required.


git-svn-id: https://svn.freebsd.org/base/head@66401 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f