MFC: Add proc/pid/stat and proc/pid/status. Make proc/cpuinfo correspond better to what Linux would show. Avoid dividing by stathz when it's 0. Disallow CREATE. PID directories are directories, not regular files.
Note that the information in proc/pid/stat and proc/pid/status is not
entirely correct (many fields are hardcoded to 0). I know how to fix most
of them, but do not have time to do so (and test the changes properly)
before 4.2, so this will have to do for now.


git-svn-id: https://svn.freebsd.org/base/stable/4@67997 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f