When to run the periodic/daily event has had several rounds of discussion over the past couple years. The most recent came to the general consensus that this was the best time, but no one actually made the change, so I'll don my asbestos undies and dive in.
Please note that this time was chosen with input from people in various
countries with various methods and schedules for switching to and from DST.
There is no perfect time to schedule this job that works for everyone, but
this time both A) Works for more people, and B) Causes problems for fewer
people. And, ultimately, you can always change it if you need to.


git-svn-id: https://svn.freebsd.org/base/head@68911 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f