Create the vn%d.ctl devices.
In a few days I will commit a patch which changes vn(4) to use the
disk-minilayer.  This will make vn(4) fully DEVFS friendly but have
the side effect that vnconfig needs the vn%d.ctl devices to be able
to configure vn(4).

Please remake your /dev/vn entries with this revision of MAKEDEV if
you don't rung DEVFS already.


git-svn-id: https://svn.freebsd.org/base/head@70101 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f