o Add cardbus support (use pci attachment for now) <bde> o Add comments in some places to clarify some points. o Don't typedef sc_p.  This isn't usually done in the drivers and may   cause problems in teh future if C goes the C++ route of requiring   one and only one definition for each and every type.  Instead use   the current convetion of expanding struct ${1}_softc * inline needed. o change some comments to be more style(9)-like. o Define and use DEV2SOFTC to encapsulate storing/getting softc from a   dev_t.  This also takes care of the missing cast from the examples. o Define and use DEVICE2SOFTC, similar to DEV2SOFTC for getting the   softc from a device_t. </bde>
We still should have this generate foo_{isa,pci,pccard,cardbus,eisa}.c
and foovar.h from templates of some sort, but I was too lazy to do
that in this commit.  I did document it in the comments, however.

Note: bde-like corrections made with the help of my my portable
plastic bde icon.  Results with the real bde may vary with use.


git-svn-id: https://svn.freebsd.org/base/head@70165 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f