Sort out the umount breakage introduced in the last 2 revisions.
Please note:
When committing changes to this file, it is important to note that
linux is not freebsd -- their system call numbers (and sometimes names)
are different on different platforms.  When in doubt (and you always need
to be) check the arch-specific unistd.h and entry.S files in the linux
kernel sources to see what the syscall numbers really are.


git-svn-id: https://svn.freebsd.org/base/head@72741 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f