Fixed two bugs with -tag lists.
If the -tag list definition didn't have a -width modifier,
the first .It call was supposed to set the width depending
on the first argument type; if it is a macro name, use the
macro's width value; otherwise, use width value of `No'.

The following two lists should produce identical output:

.Bl -tag -width Er -compact
.It Er EINVAL
invalid argument
.El
.Bl -tag -compact
.It Er EINVAL
invalid argument
.El

If the outermost -tag list definition did't have a -width
modifier, the .It elements of inner lists might not work
(producing a list where each successive element `walks' to
the right).

Example:

.Bl -tag -compact
.It outer
.It outer
.Bl -tag -compact
.It inner
.It inner
.El
.It outer
.It outer
.El

Ported from:	mdocNG


git-svn-id: https://svn.freebsd.org/base/head@73273 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f