MFC, rev 1.108:
In xpt_set_transfer_settings(), force async if either the period or the
offset is set to 0.

Re-arrange the DT limiting code so that we don't end up setting the period
to 0xa if the user really wants async.  The previous behavior seemed to
confuse the aic(4) driver.

PR:		kern/22733
Reviewed by:	gibbs
Approved by:	jkh


git-svn-id: https://svn.freebsd.org/base/stable/4@74380 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f