MFC:  RELNOTESng.
I didn't MFC the changes to the release Makefile or the nuking of the
*.TXT files.  We're not ready for that yet.  I just wanted to get the
RELNOTESng sources for RELENG_4 into the repository so I can maintain
the content in parallel with the *.TXT version of the release notes.

These files include an SGML-ified version of RELENG_4's release notes,
plus the hardware compatability lists from HEAD (minus
CURRENT-specific features such as Cardbus support).


git-svn-id: https://svn.freebsd.org/base/stable/4@77594 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f