Implement a loadable module for the oltr Olicom Token Ring driver.  I could only get a chance of testing it under 4.3, but together with the if_oltr.c fixes at least it seems to work now.  If someone has the chance to test this under -current, please do.
Unfortunaltey, the TR code itself (if_iso88025subr.c) is not written
in a way that would allow to make a seaparate KLD out of it.  By now,
just link it directly into the oltr KLD since it's probably the POLA
to be able to load the TR code together with the only TR hardware
driver we've got by now.

I've got one single unexplained panic (in doreti_switch or somewhere
there, calling a 0xc1XXXXXX address that did no longer belong to the
kernel at all) after unloading the modules once, thus i don't propose
a MFC of this module despite my testing has been done solely on 4.3,
unless someone is really going to test this stuff in -current.


git-svn-id: https://svn.freebsd.org/base/head@78218 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f