Massive MFC- probably breaking lots of rules to do this, but limited time, etc., etc...
Basic gist of changes:

Beginnings of 2300 support. Addition of a /dev entry and
ioctl entry point for the HBA (sort of meaningless in
RELENG_4- it's really only useful in DEVFS for -current-
but you *could* make the entries by hand (not worth
really asking for a reserved major as this is edge
case debugging stuff for me mostly).

Implementation of kthread for Fibre Channel loop cleanup.
Lots of buglet fixes. Ansification. Lots of target mode
cleanup.


git-svn-id: https://svn.freebsd.org/base/stable/4@79243 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f