Since revision 1.40/1.41, the default behaviour for mount_nfs is to give up after one attempt unless a background mount is requested. Background mounts would retry 10000 times (at least 7 days) before giving up.
For some situations such as diskless terminals, an NFS filesystem
may be critical to the boot process, so neither the "try once" nor
background mounts are appropiate. To cater for this situation,
unbreak the -R (retry count) parameter so that it also works in
the non-background case. Interpret a zero retry count as "retry
forever".

The defaults are now "try once" for non-background mounts and "retry
forever" for background mounts; both can be overridden via -R.
Add a description of this behaviour to the manpage.


git-svn-id: https://svn.freebsd.org/base/head@80006 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f