When -n is specified, don't attempt to turn hostnames found in utmp into addresses as we have no idea what address family they belong to.
When -n is not specified, resolve IPv6 as well as IPv4 addresses found
in the host field of utmp.  Use realhostname_sa() to resolve addresses
(the old code was wrong).

Rename ``x'' to ``x_suffix'' to avoid confusion.

Hard code the host column width to 16 (against the imminent increase
of UT_HOSTSIZE in utmp.h).


git-svn-id: https://svn.freebsd.org/base/head@80407 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f