It looks like we're doing the wrong thing by catching the setup_intr request and just calling it when we get a bridge interrupt.  The problem is that if other code wants to block hardware interrupts for a little bit with splXXX, those masks aren't updated the way we're doing it.  This doesn't matter for -current, but does for -stable.
The whole reason that we were catching interrupts was to detect that
the card was still there.  Ian's fixes however ensure that the card
will be there with an interrupt handler, or not there at all.  Since
the pcic interrupt is at a high priority, this should be OK.

This should fix the network related crashes people started seeing in
stable after I merged the pcic as a pci device code.

Submitted indirectly by: Ian Dowse
MFC when: Ian has had a chance to do his torture hang testing.


git-svn-id: https://svn.freebsd.org/base/head@82083 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f