Explain what we're doing in pcic_pci_shutdown a little better.  Try to explain the subtle side effects that are going to happen and why we go ahead and ack the interrupt source.  This stuff is tricky to get right.
Also, emperical tests have shown that doing a shutdown in attach to be
ineffectual, so remove it from there.  Analysis of the code paths
shows that nearly identical writes to these registers happen in later
parts of the code.  The hanging problem on thinkpads when we change
the interrupt routing type is something else.


git-svn-id: https://svn.freebsd.org/base/head@82282 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f