Retarget the resource parser slightly.  We only fetch current resources for the device now (we should really just be parsing a passed-in resource buffer).
Wrap long lines so this is (more) readable.

Support Address16 and Address32 resources, in the CONSUMER case.

Support DRQs so that we can handle ISA devices.

Support ExtendedIrqs (we ignore most of their attributes)

Add a placeholder device for system memory and system resources.  This
takes the place of the nexus placeholder, which only attaches to ISA.


git-svn-id: https://svn.freebsd.org/base/head@82536 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f