Move /etc/defaults/make.conf to /usr/share/examples/etc/make.conf as discussed on the arch@ mailinglist (after repo-copy).
sys.mk will .error if it finds /etc/defaults/make.conf but include
it anyways (this is the same behaviour as with the make.conf.local
removal).

/usr/share/examples/etc/make.conf has BDEFLAGS commented out now,
since it's only an example file.

Adjust all textes that talk about make.conf or defaults/make.conf to
match the new situation.


git-svn-id: https://svn.freebsd.org/base/head@82604 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f