MFC:
The timed protocol is not implemented in a compatible way by all vendors;
the size of the tsp_name field is OS-dependent.  4.3BSD used a 32-byte
field, FreeBSD uses MAXHOSTNAMELEN and RedHat apparently uses a 64-byte
field.  As a result, sanity checking code added a few months ago to detect
short packets will fail when interoperating with one of these other vendors.

Change the short packet detection code to expect a minimum packet size
corresponding to the 4.3BSD implementation, which should be a safe minimum
size.

Approved by:	jkh


git-svn-id: https://svn.freebsd.org/base/stable/4@82638 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f