Remove the panic when trying to register a sysctl with an oid too high. This stops panics on unloading modules which define their own sysctl sets.
However, this also removes the protection against somebody actually
defining a static sysctl with an oid in the range of the dynamic ones,
which would break badly if there is already a dynamic sysctl with
the requested oid.

Apparently, the algorithm for removing sysctl sets needs a bit more work.
For the present, the panic I introduced only leads to Bad Things (tm).

Submitted by:	many users of -current :(
Pointy hat to:	roam (myself) for not testing rev. 1.112 enough.


git-svn-id: https://svn.freebsd.org/base/head@84832 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f