Add the script that will be used to maintain src/share/misc/pci_vendors. The script written and used originally by msmith has been lost.
This version takes the Boemler and Heckenbach lists and produces merged
output.  It defaults to ignoring any entries from Heckenbach already
found in Boemler but the -l option causes it to take the entry with the
longest description where an entry appears in both lists.

If this script is replaced, care should be taken to

1) Always use upper-case hexidecimal tokens in device ids.

2) Always keep device lists sorted within vendor lists, which must also
   be sorted.

3) Do not try to include input from the previous pci_vendors file, since
   bogus ids seem to be removed from both the Boemler and Heckenbach
   lists from time to time.


git-svn-id: https://svn.freebsd.org/base/head@86781 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f