Regen.
This file is now generated using src/tools/tools/pciid/mk_pci_vendors.pl,
which merges the Boemler and Heckenbach lists used for rev 1.2.

For now, mk_pci_vendors.pl is called with the -l option, which uses
the entry with the longer description where the same device or vendor
is found in both lists.

If it turns out that this causes to much "back-and-forth" in future
deltas, we can drop the use of the -l option.


git-svn-id: https://svn.freebsd.org/base/head@86813 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f