Don't adjust_linklocal() when pulling a sockaddr out of an ncpaddr or ncprange structure.
Don't write() the netmask for IPv6 sockaddrs to the routing socket if
the prefixlen is 128.

It seems that messages written to the routing socket with the scopeid
set for link local addresses are not understood.  Instead, we have to
put the scopeid in the 5th and 6th bytes of the address (see
adjust_linklocal() in ncpaddr.c).  I think this may be a bug in the
KAME implementation - it should really understand both forms.


git-svn-id: https://svn.freebsd.org/base/head@86815 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f