Update the default newfs block and fragment sizes from 8192/1024 to 16384/2048.
Following recent discussions on the -arch mailing list, involving dillon
and mckusick, this change parallels the one made over a decade ago when
the default was bumped up from 4096/512.

This should provide significant performance improvements for most
folks, less significant performance losses for a few folks and
wasted space lost to large fragments for many folks.

For discussion, please see the following thread in the -arch archive:

Subject: Using a larger block size on large filesystems

The discussion ceases to be relevant when the issue of partitioning
schemes is raised.


git-svn-id: https://svn.freebsd.org/base/head@87661 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f