Fix the "conexant chips don't work in full duplexmode" problem. According to Phil Kernick:
"The problem is that in full duplex mode, the Conexant chip always reports a
carrier lost error, even when the frame is successfully sent.  So, if we
have a Conexant chip, then ignore carrier lost when in full duplex
mode."

Since the Xircom chips seem to have the same issue and since we already
have a workaround for this, just expand the workaround test to also
check for DC_IS_CONEXANT().


git-svn-id: https://svn.freebsd.org/base/head@88214 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f