Re-introduce the fix that delays mounting of network filesystems until the network is initialized.  This was first implemented in rev 1.268 of src/etc/rc, but was backed out at wollman's request.
The objection was that the right place for the fix is in mount(8).
Having looked at that problem, I find it hard to believe that
the hoops one would have to jump through can be justified by the
desire for purity alone.

Note that there are reported issues surrounding nfsclient kernel
support and mount_nfs(8), which currently make NFS an ugly exception
to the general case.

With this change, systems with non-NFS network filesystems configured
for mounting on startup in /etc/fstab are no longer guaranteed to
fail on startup.


git-svn-id: https://svn.freebsd.org/base/head@88531 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f