Add 'camcontrol rescan all' and 'camcontrol reset all' functionality to camcontrol.
This enables rescanning all busses or resetting all busses in a system.
The current implementation is not the ideal way to do it -- the ideal way
to do it would be for the transport layer to handle wildcarded busses on
bus rescan and reset operations.  The current implementation enumerates all
the busses and sends a rescan or reset CCB individually.  Handling this
behavior in the transport layer will happen later.

Reviewed by:	imp
Tested by:	joerg
MFC after:	1 week


git-svn-id: https://svn.freebsd.org/base/head@89515 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f