Readded the svc_create() and the registering of the local transport, now it is fixed. This should get us a working keyserv again, since it depends on local transport for key exchange.
Since we do not have any KEYFILE name hardcoded anymore, set the
umask that way that the keyserver socket can be created with with
the appropriate permissions.

Re-add the accidently removed signal(SIGPIPE, SIG_IGN); to the code
which makes sense to avoid SIGPIPE when a disconnect on rpc socket
occurs.

Submitted by: mbr


git-svn-id: https://svn.freebsd.org/base/head@90317 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f