Move the declaration of panic() from sys/param.h back to sys/systm.h.
Almost all .c files have to include <sys/systm.h> for more than its
declaration of panic(), so little is gained from declaring panic() in
a wrong place.  This probably depends on missing garbage collection
of the includes of <sys/systm.h> that were added to get snprintf()
declared for old versions of the ktr macros.


git-svn-id: https://svn.freebsd.org/base/head@90503 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f