make(1) claims to ignore the remainder of a conditional once its value has been determined similar to C. That is, one expects a construction like,
.if defined(TEST) && (${TEST:L} == "test")

Never to generate an error since the second expression should never be
evaluated when TEST is undefined.

However, this was not the case. The above fails with the current
make(1) if TEST is undefined. This patch fixes the above and many
similar cases.

PR:		bin/34032
Submitted by:	Alan Eldridge <alane@geeksrus.net>
MFC after:	1 week


git-svn-id: https://svn.freebsd.org/base/head@90965 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f