Add a new test_counter() function which tries to determine the width of the inter-value histogram for 2000 samples.  If the width is 3 or less for 10 consequtive samples, we trust the counter to be good, otherwise we use the *_safe() method.
This method may be too strict, but the worst which can happen is that
we take the performance hit of the *_safe() method when we should not.

Make the *_safe() method more discriminating by mandating that the three
samples do not span more than 15 ticks on the counter.

Disable the PCI-ident based probing as a means to recognize good
counters.

Inspiration from:	dillon and msmith


git-svn-id: https://svn.freebsd.org/base/head@91237 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f