Don't deref NULL mutex pointer when pipeclose()'ing a pipe that is not fully instaniated.
Revert the logic in pipeclose so that we don't have the entire function
pretty much under a single if() statement, instead invert the test and
just return if it fails.

Submitted (in different form) by: bde

Don't use pool mutexes for pipes.  We can not use pool mutexes
because we will need to grab the select lock while holding a pipe
lock which is not allowed because you may not aquire additional
mutexes when holding a pool mutex.

Instead malloc(9) space for the mutex that is shared between the
pipes.


git-svn-id: https://svn.freebsd.org/base/head@91968 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f