As a XXX bandaid open the mounted device READ/WRITE even if we only mount read-only.
The trouble here is that we don't reopen the device in read/write mode
when we remount in read/write mode resulting in a filesystem sending
write requests to a device which was only opened read/only.

I'm not quite sure how such a reopen would best be done and defer
the problem to more agile hackers.


git-svn-id: https://svn.freebsd.org/base/head@92092 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f