Improve support of MFC cards (Multi-function cards).  This commit allows us to properly parse cards with attribute memory based CIS that before wouldn't parse correctly, sometimes with a panic.  This allows me to get my 3C562 modem/ethernet card to fail to attach due to problems in the ep and sio drivers rather than due to problems in the CIS parsing code :-).
We weren't setting the address to jump to for the function entries.
This caused us to only work when the addional entries were after the
first ones.  On the 3C562/3C563 card this was not the case.

We were also mapping Attribute memory when common memory was asked for
in the target of the LONGLINK_{A,C} or LONGLINK_MFC.

My IBM Home And Away Modem/LAN card still fails for reasons unknown.


git-svn-id: https://svn.freebsd.org/base/head@93370 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f