MFC 1.308
brelse() was improperly clearing B_DELWRI in the B_DELWRI|B_INVAL case
without removing the buffer from the vnode's dirty buffer list, which
can result in a panic in NFS.  Replaced the code with a call to bundirty()
which deals with it properly.

PR:             kern/36108, kern/36174
Special mention: to Danny Schales <dan@coes.LaTech.edu> for providing a core dum
p that helped me track this down.


git-svn-id: https://svn.freebsd.org/base/stable/4@93729 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f