Despite the fact that it is documented in the handbook, the release notes, and UPDATING and has been posted to both freebsd-current and freebsd-stable, users are still not adding the required smmsp user and group before doing an installworld.  Therefore, don't let users do an installworld unless they have followed directions.
Add a new installcheck Makefile target which installworld runs before
actually starting the installation.  This target can be used by other parts
of userland as well.  The first addition to the target is to check for the
smmsp user and group if NO_SENDMAIL isn't defined.

Others may add checks to this target as they see fit.

MFC after:	1 week


git-svn-id: https://svn.freebsd.org/base/head@95146 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f