Remove the temporary alignment check in free().
Implement the following checks on freed memory in the bucket path:
	- Slab membership
	- Alignment
	- Duplicate free

This previously was only done if we skipped the buckets.  This code will slow
down INVARIANTS a bit, but it is smp safe.  The checks were moved out of the
normal path and into hooks supplied in uma_dbg.


git-svn-id: https://svn.freebsd.org/base/head@95899 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f