MFC:
If we get a DATA UNDERRUN error from QLogic FC cards, but the RQCS_RU bit
is not set in the scsi completion status, or if the residual is clearly
nonsense, then this was a command that suffered the loss of one or more
FC frames in the middle of the exchange.

Set HBA_BOTCH and hope it will get retried. It's the only thing we can do.

Approved by: re@freebsd.org


git-svn-id: https://svn.freebsd.org/base/stable/4@95907 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f