Don't send a RAD_NAS_IP_ADDRESS attribute as RAD_NAS_IDENTIFIER is sufficient.
In fact, using both breaks the radiator RADIUS daemon when used with
a db as it maps both attributes to the same field value and then
fails the insert.

I decided to remove RAD_NAS_IP_ADDRESS on the basis that rfc2138 says:

      An Access-Request MUST contain a User-Name attribute.  It SHOULD
      contain either a NAS-IP-Address attribute or NAS-Identifier
      attribute (or both, although that is not recommended).  It MUST

despite the fact that this not recommended bit was removed from the
updated rfc.


git-svn-id: https://svn.freebsd.org/base/head@97738 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f