MFC *: Almost complete rewrite of the lpc commands 'abort', 'enable', 'disable', 'restart', 'start', 'stop' and 'up'.  This reduces the amount of code that is run as the priv userid, and eliminates some cases where error messages were written while that priv uid was in effect.
As far as users are concerned, there should be no noticable difference
in the new versions.  In case there *is*, the previous implementations
are still there as 'xabort', 'xenable', etc, so they are available for
instant fallback.  If no one reports a problem after a few weeks, then
a later update will remove those x-commands.

[* = 1.26->1.28 common.c, 1.21 lp.h, 1.25 cmds.c, 1.5 cmdtab.c, 1.6 extern.h]


git-svn-id: https://svn.freebsd.org/base/stable/4@98645 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f