MFC compile.c::
revision 1.21
date: 2002/06/01 13:25:47;  author: green;  state: Exp;  lines: +24 -4
Fix a bug in sed(1)'s "s" command wherein if an escape ("\" character)
was initiated at the last character of the line buffer, the Wrong
Thing was done and sed barfed by interpreting the following NUL byte
as a digit.  Instead, pull up the next buffer and record that the "\"
was last seen.

Submitted by:	green


git-svn-id: https://svn.freebsd.org/base/stable/4@98814 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f