Fix a bazillion lint and WARNS warnings. One major fix is the removal of semicolons from the end of macros:
#define FOO() bar(a,b,c);

becomes

#define FOO() bar(a,b,c)

Thus requiring the semicolon in the invocation of FOO. This is much
cleaner syntax and more consistent with expectations when writing
function-like things in source.

With both peril-sensitive sunglasses and flame-proof undies on, tighten
up some types, and work around some warnings generated by this. There
are some _horrible_ const/non-const issues in this code.


git-svn-id: https://svn.freebsd.org/base/head@100113 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f