Introduce support for Mandatory Access Control and extensible kernel access control.
Invoke an appropriate MAC entry point to authorize execution of
a file by a process.  The check is placed slightly differently
than it appears in the trustedbsd_mac tree so that it prevents
a little more information leakage about the target of the execve()
operation.

Obtained from:	TrustedBSD Project
Sponsored by:	DARPA, NAI Labs


git-svn-id: https://svn.freebsd.org/base/head@101158 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f