Remember to unlock the (optional) vnode in vfs_stdextattrctl().  Failing to do this made the following script hang:
	#!/bin/sh
	set -ex

	extattrctl start /tmp
	extattrctl initattr 64 /tmp/EA00
	extattrctl enable /tmp user ea00 /tmp/EA00
	extattrctl showattr /tmp/EA00

if the filesystem backing /tmp did not support EAs.

The real solution is probably to have the extattrctl syscall do the
unlocking rather than depend on the filesystem to do it.  Considering
that extattrctl is going to be made obsolete anyway, this has dogwash
priority.

Sponsored by:	DARPA & NAI Labs.


git-svn-id: https://svn.freebsd.org/base/head@101786 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f