Synchronize ipfw2 with the version in -current (adding sets of rules, prevention of loops in keepalive generation, better defaults on size of dynamic rule table).
For documentation, please refer to the ipfw manpage in -current
(which I am going to MFC as soon as I have completed the
section listing differences between ipfw1-stable and ipfw2).
In particular have a look at the sections "PACKET FLOW",
"IPFW2 ENHANCEMENTS" and "EXAMPLES" to see if your ruleset
can be simplified with the new commands.


git-svn-id: https://svn.freebsd.org/base/stable/4@101980 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f