When login tries to do the chmod/chflags on a read only file system, it complains that it can't do it because the filesystem is readonly. Assume that when the user has a readonly /dev that they don't care if login can't change the permissions/flags.  While this does break a few things like msgs, we'll assume that the user setting up the read only system knows what they are doing.
All this change does is to stop the complaint when the file system is
read only.  It also adds comments as to why EROFS and EOPNOTSUPP are
ignored.

This allows one to have a read-only / w/o a /dev MFS and have a
relatively warning-free existence.  /etc/rc still complains when it
can't chown/chflags/chmod things, but that's easy to ignore/tweak.

Reviewed by: roberto, phk
Sponsored by: Timing Solutions


git-svn-id: https://svn.freebsd.org/base/head@102141 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f