Convert to new assembler field syntax.
Document the SXFRCTL2 register found on U2 and U160 controllers.

Overload the MWI_RESIDUAL field for use as the SCB to be downloaded
for "immediate" (or those without the disconnect privledge)
transactions.

Add scratch ram locations for the 274X that give us a bit more
information including whether to enable extended translation.


git-svn-id: https://svn.freebsd.org/base/head@102672 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f