MFC:
a) we don't believe what the board tells us all the time (if the BIOS
hasn't run, port page 2 and port page 0 tend to be garbage)

b) add the missing code to set parameters for the SPI cards.

c) Fix copyrights in mpilib in concordance with LSI and other
freebsd developers.


git-svn-id: https://svn.freebsd.org/base/stable/4@102823 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f