Introduce to the !program specification a capability similar to the one that already exists for hosts: being able to specify a section that applies to every program *except* the one in question.
The normal syntax for program specification is still valid. For the new
capability, one uses:

!-program

Since there is no way to specify a program beginning with a dash in the old
syntax, as it would be interpreted as the case above, the following
alternative syntax to the original capability is provided:

!+program

This shouldn't introduce incompatibilities with any syslogd configuration
in production because -stable's syslogd does not support a dash anywhere in
the program specification.

MFC after:	2 weeks


git-svn-id: https://svn.freebsd.org/base/head@103855 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f