Implement DCE 1.1 compliant UUID functions. Immediate use of these functions is expected for uuidgen(1), mca(8) and gpt(8). Given the generic use of UUIDs beyond the scope of the DCE 1.1 specification, visibility of the data structure at all levels of the machine, including firmware and the wish to not create a permanent build- time FreeBSD-ism for DCE compliant applications by creating a new library, it was decided that libc would be the least inappropriate place. Also, because the UUID functions live in libc under IRIX as well, we have maximized our portability and left as many options open as possible.
This implementation introduces an extension not found in the
specification: the status parameter is allowed to be a NULL-
pointer. The reason for introducing the extension is because
the status is almost never of any use.

The manpage that's part of this commit is a minimal place-holder
and is further fleshed-out in the near future.

Approved by: re@
Contributed by: Hiten Mahesh Pandya <hiten@unixdaemons.com>
Sponsored by: marcel :-)
Tested on: alpha, i386, ia64


git-svn-id: https://svn.freebsd.org/base/head@106184 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f