Unifdef has got some crack-smoking exit values:    "The unifdef utility exits 0 if the output is an exact copy of the input,     1 if not, and 2 if in trouble."
This causes an 'Error code 1 (ignored)' in the world output, which upsets
the whereintheworld scripts that the tinderboxes use.

However, this stuff here is a relic.  We do not enable HAVE_TCL_INTERP
nor HAVE_PERL_INTERP..  The Makefile hooks to turn them on have gone ages
ago, and this stuff was here for release building purposes only.  Rather
than fight with the tinderbox builds, clean house a bit and remove the
last remaining unused relics of this stuff.


git-svn-id: https://svn.freebsd.org/base/head@108679 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f