Pave the road to removing the fixed size limit on device nodes:
Change the si_name of dev_t's to be a char * and put a private buffer for
holding the name at then end of the struct.

Initialize si_name to point to the private buffer.

Put a KASSERT in geom_disk to prevent overrun on the fake dev_t we still
have to generate for the disk_drivers.


git-svn-id: https://svn.freebsd.org/base/head@110317 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f