Let tilde expansion be done even if a file/directory doesn't exist yet. This makes such natural commands as "MKD ~user/newdir" or "STOR ~/newfile" do what they are supposed to instead of failing miserably with the "File not found" error.
This involves a bit of code reorganization.  Namely, the code doing
glob(3) expansion has been separated to a function; a new function
has been introduced to do tilde expansion; the latter function is
invoked on a pathname before the former one.  Thus behaviour mimicing
that of the Bourne shell has been achieved.


git-svn-id: https://svn.freebsd.org/base/head@110340 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f