MFC: IPSEC_FILTERGIF config option
Add a new config option IPSEC_FILTERGIF to control whether or not
packets coming out of a GIF tunnel are re-processed by ipfw, et. al.
By default they are not reprocessed.  With the option they are.

This reverts 1.214.  Prior to that change packets were not re-processed.
After they were which caused problems because packets do not have
distinguishing characteristics (like a special network if) that allows
them to be filtered specially.

PR:		48159
Reviewed by:	Guido van Rooij <guido@gvr.org>
Approved by:	re (jhb, murray)


git-svn-id: https://svn.freebsd.org/base/stable/4@111328 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f