If devstat_new_entry() is passed a unit number of -1 assume that the devstat is for an "interior" GEOM node and register using the name argument as a geom identity pointer.  Do not put these devstat structures on the list returned by the sysctl.
This gives us the ability to tell the two kinds of nodes apart and
leave the current "strictly physical" view of devstat intact without
modifications, yet be able to use devstat for both kinds of devices.

It also saves us bloating struct devstat with another 48 bytes of
space for the name.  At least for now.

Reviewed by:    ken


git-svn-id: https://svn.freebsd.org/base/head@112369 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f