Attempt to automatically read in kernel module symbols when a live or dead kernel core is loaded into gdb. This extends gdb's existing shared library support, so the "info sharedlibrary", "sharedlibrary" and "nosharedlibrary" commands can be used to view and change the list of loaded symbol files.
The current implementation is more than a kludge however, and it
will not always manage to find the .ko.debug file corresponding to
the loaded module. In particular, for modules whose build directory
cannot be easily guessed from the module name such as all the
netgraph modules, the debug version of the .ko will not be found
automatically.

The logic for finding the module file first attempts to guess at
the module build directory by parsing the version[] string. Then
using that directory ($DIR), it tries the following paths in turn:

	./<module>.ko.debug		./<module>.ko
	$DIR/<module>.ko.debug		$DIR/<module>.ko
	/boot/kernel/<module>.ko.debug	/boot/kernel/<module>.ko

Approved by:	obrien, mp


git-svn-id: https://svn.freebsd.org/base/head@112454 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f