Commit a partial lazy thread switch mechanism for i386.  it isn't as lazy as it could be and can do with some more cleanup.  Currently its under options LAZY_SWITCH.  What this does is avoid %cr3 reloads for short context switches that do not involve another user process.  ie: we can take an interrupt, switch to a kthread and return to the user without explicitly flushing the tlb.  However, this isn't as exciting as it could be, the interrupt overhead is still high and too much blocks on Giant still.  There are some debug sysctls, for stats and for an on/off switch.
The main problem with doing this has been "what if the process that you're
running on exits while we're borrowing its address space?" - in this case
we use an IPI to give it a kick when we're about to reclaim the pmap.

Its not compiled in unless you add the LAZY_SWITCH option.  I want to fix a
few more things and get some more feedback before turning it on by default.

This is NOT a replacement for Bosko's lazy interrupt stuff.  This was more
meant for the kthread case, while his was for interrupts.  Mine helps a
little for interrupts, but his helps a lot more.

The stats are enabled with options SWTCH_OPTIM_STATS - this has been a
pseudo-option for years, I just added a bunch of stuff to it.

One non-trivial change was to select a new thread before calling
cpu_switch() in the first place.  This allows us to catch the silly
case of doing a cpu_switch() to the current process.  This happens
uncomfortably often.  This simplifies a bit of the asm code in cpu_switch
(no longer have to call choosethread() in the middle).  This has been
implemented on i386 and (thanks to jake) sparc64.  The others will come
soon.  This is actually seperate to the lazy switch stuff.

Glanced at by:  jake, jhb


git-svn-id: https://svn.freebsd.org/base/head@112993 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f