Add a couple asserts to pthread_cond_foo to ensure the (low-level) lock level is 0.  Thus far, the threads implementation doesn't use mutexes or condition variables so the lock level should be 0.
Save the return value when trying to schedule a new thread and
use this to return an error from pthread_create().

Change the max sleep time for an idle KSE to 1 minute from 2 minutes.

Maintain a count of the number of KSEs within a KSEG.

With these changes scope system threads seem to work, but heavy
use of them crash the kernel (supposedly VM bugs).


git-svn-id: https://svn.freebsd.org/base/head@113870 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f