Work around SCSI spec violation by the Quantum Atlas 10K. This drive delays going async after receiving a WDTR message.  We now send an SDTR message after a WDTR even if our goal is to go async.  This should work even for confused devices.
If we get an unexpected busfree when attempting a WDTR
or SDTR, only set the goal negotiation parameters we were
trying to negotiate to off.  This means that should a WDTR
message fail, we will still try an SDTR if our goal is
non-async.

Fix a few more places where we were looking at goal.period
instead of goal.offset for determining if we should be
negotiating sync.  This should not have any impact on
our behavior, but the offset is more definitive and should
be used.


git-svn-id: https://svn.freebsd.org/base/head@115918 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f