- Add a software watchdog facility.
This commit has two pieces. One half is the watchdog kernel code which lives
primarily in hardclock() in sys/kern/kern_clock.c. The other half is a userland
daemon which, when run, will keep the watchdog from firing while the userland
is intact and functioning.

Approved by:	jeff (mentor)


git-svn-id: https://svn.freebsd.org/base/head@116874 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f