From the PR:
I am the maintainer of CTM.  There is a problem that when very large deltas
are created, that the program ctm_smail, which is responsible for mailing
the deltas out, will instead create a single message that says the delta
is too large.  However, if the -q option is set, instead of placing this
message in the queue (as it would have done with the deltas), it mails it
out directly.  This conflicts with the current working of CTM in that the
email address is set as %%REPLACE-ME%% so that the created mailing pieces
can be signed by gnu-pgp, and then have the mailing address changed.

This fix means that if the -q option is set, and the delta is too large,
the "too large" message is placed in the queue.

Also, I made the "too large" message a little more up to date.

Submitted by:	Stephen Montgomery-Smith <stephen@math.missouri.edu>
PR:		bin/50328
MFC After:	2 weeks


git-svn-id: https://svn.freebsd.org/base/head@117503 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f