Add in the promised cluster refcount limit hack.  (Not applicable to -current, which has int size refcounts.)
4.x uses char refcounts, which can be overflowed, causing neat panics.
This patch works around the problem by hooking into every m_* function
that increases the refcount and doing two things:

1.  Panicing if the refcount went negative (perhaps due to some external
function messing with the refcount.)

2.  Making a full copy of the resulting chain in order to keep the
refcount under a reasonable threshold if necessary.

The kern.ipc.m_clreflimithits sysctl has been made available so that
the number of times a copy had to be performed can be monitored.  Luckily,
this should not be a common occurance.


git-svn-id: https://svn.freebsd.org/base/stable/4@117818 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f