Introduce two new MAC Framework and MAC policy entry points:
  mac_reflect_mbuf_icmp()
  mac_reflect_mbuf_tcp()

These entry points permit MAC policies to do "update in place"
changes to the labels on ICMP and TCP mbuf headers when an ICMP or
TCP response is generated to a packet outside of the context of
an existing socket.  For example, in respond to a ping or a RST
packet to a SYN on a closed port.

Obtained from:	TrustedBSD Project
Sponsored by:	DARPA, Network Associates Laboratories


git-svn-id: https://svn.freebsd.org/base/head@119245 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f