MFC:	rev. 1.63 and rev. 1.123
Do it right this time. All davicom cards need DC_TX_ALIGN.

Avoid an unnecessary copy of a packet if it is already in a single mbuf.
Introduce an additional device flag for those NICs which require the
transmit buffers to be aligned to 32-bit boundaries.

Approved by:		murray (re)


git-svn-id: https://svn.freebsd.org/base/stable/4@119503 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f