Revert rev. 1.378, and restore the correct arithmetic expression syntax.  The
	make buildworld
	mv /usr/include /usr/include.old
	make installworld

issue has been fixed a month ago in Makefile,v 1.285, and there
is no valid reason to continue to keep the wrong syntax here --
buildworld takes care of upgrading a make for you if necessary.
But if you find yourself in an environment with an old make(1)
binary that breaks on this, and this is because you attempted
to run a target other than buildworld, don't whine but try again
with -DALWAYS_CHECK_MAKE defined -- it should do the trick.
Otherwise, if you still have a problem, please report it as a
bug and attach the ``make -dl ...'' output.

Reviewed by:	marcel


git-svn-id: https://svn.freebsd.org/base/head@119580 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f