In re_diag(), there's no need for us to call re_start() ourselves: IF_HANDOFF() does it for us behind the scenes. Remove the extra call to re_start() otherwise we try to transmit twice.
In re_encap(), fix the code that guards against consuming too many
descriptors in the TX ring so that it actually works. With the
new 8169S chip, I was able to hit a corner case that drained the
free descriptor count all the way to 0. This is not supposed to
be possible.


git-svn-id: https://svn.freebsd.org/base/head@120208 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f