The present defaults for the open and close for device drivers which provide no methods does not make any sense, and is not used by any driver.
It is a pretty hard to come up with even a theoretical concept of
a device driver which would always fail open and close with ENODEV.

Change the defaults to be nullopen() and nullclose() which simply
does nothing.

Remove explicit initializations to these from the drivers which
already used them.


git-svn-id: https://svn.freebsd.org/base/head@120506 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f