Fix a bug which was introduced in version 1.187.2.16 of sys/vm/vm_map.c and affects the unwiring memory:
The map entries are processed in a loop, checking to make sure the
entry is wired and asserting it has a wired count. However, another
loop was inserted more-or-less in the middle of the of the unwiring
path.  This loop picks up the "entry" loop variable from the first loop
without first setting it to start_entry.  Naturally, the second loop
is never entered and the pages backing the entries are never unwired.
This can lead to a leak of wired pages.

Reviewed by: alc
Approved by: re (murray)


git-svn-id: https://svn.freebsd.org/base/stable/4@120618 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f