When we cannot allocate an external buffer (bacause we've hit the maximum number of pages for buffers) return -1 instead of 0. This fixes a panic under conditions when many mbufs are needed.
Update the head pointer of the receive buffer pool queue even when
we could not supply a buffer to the chip. Otherwise the chip will
not re-interrupt us for another try. A better strategy would probably
be to remember this condition and to supply buffers without an interrupt
as soon as buffers get available.


git-svn-id: https://svn.freebsd.org/base/head@121617 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f