Update the Stallion driver to 5.6.0b1 (the latest one on the web site).  This only updates the stallion (stl) driver.  The stli driver isn't updated because the driver is corrupt on the stallion web site :-(.
This fixes my EC8/32-AT card, which worked with the 1.0.0 driver, but
not the 2.0.0 driver.  It also fails to break my EasyIO-8 PCI card
(well, I can't completely test it just yet).

This is being committed directly to stable because this driver doesn't
compile in current.  I'm doing a current port, but it will commit that
later.  Also, 2.0.0 hasn't made it into current either.


git-svn-id: https://svn.freebsd.org/base/stable/4@123335 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f