Overhaul of the timeout/reinit framework. This should clear up most of the leftovers from the old version that really doesn't work anymore.
Add a reset function for host-end of the ATA channel. This is needed
for the SiI3112 in order to whack it back to reality if a device
locks up the SATA interface (thereby preventing that we can reset the
device). The result is that ATA now recovers from the timeouts that
happens with the SiI3112A and more or less all disks based on old
PATA electronics with a Marvell PATA->SATA converter. This includes
lots of the popular SATA dongles and the WDC Raptor disks..


git-svn-id: https://svn.freebsd.org/base/head@124403 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f