Actually program the list of recording devices in sv_mix_setrecsrc(). This change has not been tested.
This change was triggered by a gcc(1) warning on ia64 at -O2.  The
variable v was not used after being computed, which resulted in enough
dead code elimination (DCE) to confuse the compiler and emit a bogus
warning about the use of the variable i without prior definition. The
variable i is the loop variable.

Submitted by: des
Responsibility: marcel


git-svn-id: https://svn.freebsd.org/base/head@127218 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f