Insert the directories passed with the T134 bootp cookie (kern.bootp_cookie) in the list of paths used for diskless templates.
When foo.cpio.gz archives are present, do not copy the content from a
foo/ directory even if present. This restores the previous behaviour,
and is also important from the point of view of efficiency (saving
many RPCs to copy the individual files), which was the whole point for
introducing the cpio archive.

(these files are not anymore in -current but equivalent functionality
will be implemented there, and the maintainer has been already informed).


git-svn-id: https://svn.freebsd.org/base/stable/4@127368 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f