Associate a simple count of waiters with each condition variable.  The count is protected by the mutex that protects the condition, so the count does not require any extra locking or atomic operations.  It serves as an optimization to avoid calling into the sleepqueue code at all if there are no waiters.
Note that the count can get temporarily out of sync when threads sleeping
on a condition variable time out or are aborted.  However, it doesn't hurt
to call the sleepqueue code for either a signal or a broadcast when there
are no waiters, and the count is never out of sync in the opposite
direction unless we have more than INT_MAX sleeping threads.


git-svn-id: https://svn.freebsd.org/base/head@127954 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f