Adjustments to package sets for disc1 for the 4.10 release.  The compat packages are being removed because of security issues, they are no longer buildable.  The other packages being removed are due to space issues, they don't fit.  Packages chosen for removal were either thought to be less popular with the majority of people, or were large *and* you need a network connection to be able to use them so downloading them post-install should be possible.
This will be merged into RELENG_4_10 immediately but note that a commit
to HEAD has not and will not be done for this now.  -current package
sets are smaller (.tbz versus .tgz) so the package sets are at least
somewhat disjoint.  And it's pointless to update this in HEAD until
a 5.X release comes.


git-svn-id: https://svn.freebsd.org/base/stable/4@128853 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f