MFC *:    Fix the kludge-old-options processing so `ps tpt' will be treated the same as `ps -tpt', instead of being changed into `ps -tpT'.  Also adjust it so `ps t p0' is treated like `ps -t p0', instead of making it `ps -T p0'.
This includes the rewrite of the kludge-option processing, but it does NOT
remove the long-undocumented BACKWARD_COMPATIBILITY compile-time option.
If you delete the '#define BACKWARD_COMPATIBILITY' line, you'll get the
newer handling for "extra arguments", which is to treat them as requests
for processes.

Also brings in a number of style(9) improvements.

[* = 1.90 & 1.92 -> 1.98 ps.c]

PR:		bin/52489         (for `ps tpt' and `ps t p0')
Submitted by:	Jilles Tjoelker   ( " )


git-svn-id: https://svn.freebsd.org/base/stable/4@130220 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f