FreeBSD-ify ALTQ:  - add locking  - disable ALTQ3_COMPAT by default (do not remove the code to keep the diff    towards KAME small)  - put some more code under ALTQ3 conditional compilation as it should be  - account for if_xname  - some more minor compile fixes
As people started wondering:
The strange path layout "altq/altq" is there to avoid "-Isys/contrib" and
make it "-Isys/contrib/altq" instead, as we will need at least <altq/altq.h>
and <altq/if_altq.h> for kernel compilation.

The "freebsd4_..." in the privious commit is just the best tag name in the
KAME tree I could find to classify this in order to track its history. It
does *not* mean that this will go to 4-STABLE or anything of that kind.


git-svn-id: https://svn.freebsd.org/base/head@130368 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f