Try harder to give new processes a clean initial fpu state.  fpu_cleanstate wasn't actually clean, it was saving the xmm registers as left over by the bios.  fninit() doesn't clear those.
In fpudna(), instead of doing a fninit() and forgetting to load the initial
mxcsr, do a full fxrstor(&fpu_cleanstate).  Otherwise we hand over whatever
random values are left in the xmm registers by the last user.

I'm not certain of whether this is excessive paranoia or not, but there was
an outright bug in neglecting to set the mxcsr value that caused awk to
SIGFPE in some case.  Especially for Tim Robbins. :-)

i386 probably should do something about the mxcsr setings too.

Found by:  tjr


git-svn-id: https://svn.freebsd.org/base/head@130667 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f