Removed foot-shooting setting of CR0_TS in exec_setregs().  It is unnecessary because cpu_setregs() and/or npxinit() always sets CR0_TS during system initialization, and CR0_TS is set in the next statement (fpstate_drop()) if necessary after system initialization.  Setting it unnecessarily was less than a pessimization since it broke the invariant that the npx can be used without an npxdna() trap if fpucurthread is non-null.  The broken invariant became harmful when I added an fnclex to npxdrop().
Removed setting of CR0_MP in exec_setregs().  This was similarly
unnecessary but was harmless.

Updated comments (mainly by removing them).  Things are simpler now
that we have cpu_setregs() and don't support a math emulator or pretend
to support not having either a math emulator or an npx.

Removed the ifdef for avoiding setting CR0_NE in the !SMP case in
cpu_setregs().  npx_probe() should reverse the setting if it wants to
force IRQ13 exception handling for testing.


git-svn-id: https://svn.freebsd.org/base/head@130722 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f