Stop pretending we support full-blown HTTP/1.1 because we are not ready to handle all its aspects.  Failure to deal with HTTP/1.1, with chunked transfer encoding in particular, led to broken transfers and data corruption.  Fall back to using HTTP/1.0, which is safe.
A point to note: While the "Host:" header had not been in the
original HTTP/1.0 standard, it became a de-facto standard for
HTTP/1.0 as well.  It is understood by servers in HTTP/1.0
sessions if supported at all.

This issue doesn't apply to CURRENT since lukemftp handles HTTP/1.1
correctly.

PR:		bin/68547
Reviewed by:	vs


git-svn-id: https://svn.freebsd.org/base/stable/4@132252 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f