Move the initialization and teardown of pmaps to the vmspace zone's init and fini handlers.  Our vm system removes all userland mappings at exit prior to calling pmap_release.  It just so happens that we might as well reuse the pmap for the next process since the userland slate has already been wiped clean.
However.  There is a functional benefit to this as well.  For platforms
that share userland and kernel context in the same pmap, it means that
the kernel portion of a pmap remains valid after the vmspace has been
freed (process exit) and while it is in uma's cache.  This is significant
for i386 SMP systems with kernel context borrowing because it avoids
a LOT of IPIs from the pmap_lazyfix() cleanup in the usual case.

Tested on:  amd64, i386, sparc64, alpha
Glanced at by:  alc


git-svn-id: https://svn.freebsd.org/base/head@132475 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f