Fix a race in vm_page_sleep_if_busy().  Due to vm_object locking being incomplete, it currently has to know how to drop and pick back up the vm_object's mutex if it has to sleep and drop the page queue mutex.  The problem with this is that if the page is busy, while we are sleeping, the page can be freed and object disappear.  When trying to lock m->object, we'd get a stale or NULL pointer and crash.
The object is now cached, but this makes the assumption that
the object is referenced in some manner and will not itself
disappear while it is unlocked.  Since this only happens if
the object is locked, I had to remove an assumption earlier in
contigmalloc() that reversed the order of locking the object and
doing vm_page_sleep_if_busy(), not the normal order.


git-svn-id: https://svn.freebsd.org/base/head@132517 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f