MFC: Enable KEYMAP_PATH environment variable to tweak keymap path.
This commit imports following commit to RELENG_4:
	src/usr.sbin/kbdcontrol/kbdcontrol.1 rev 1.27
	src/usr.sbin/kbdcontrol/kbdcontrol.c rev 1.34, 1.35

If my build sandbox works as expected, Jul/17/2004 or later RELENG_4
release build is broken iff old build environment is used.  This is because
that even if a new keymap hits the tree, release build doesn't see src/
directory but /usr/share/syscons/keymaps.

About three years before, 5-current had a fix for that by sobomax, but
it was not MFCed yet.  Let me give a chance to fix for RELENG_4.

Tested on: tora.jp.FreeBSD.org
           (aka snapshots.jp.FreeBSD.org RELENG_4 buildbox)


git-svn-id: https://svn.freebsd.org/base/stable/4@133076 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f