Switch over to a different, more flexible test output protocol that's understood by Perl's Test::Harness module and prove(1) commands.
Update README to describe the new protocol.  The work's broken down into
two main sets of changes.

First, update the existing test programs (shell scripts and C programs)
to produce output in the ok/not ok format, and to, where possible, also
produce a header describing the number of tests that are expected to be
run.

Second, provide the .t files that actually run the tests.  In some cases
these are copies of, or very similar too, scripts that already existed.
I've kept the old scripts around so that it's possible to verify that
behaviour under this new system (in terms of whether or not a test fails)
is identical to the behaviour under the old system.

Add a TODO file.


git-svn-id: https://svn.freebsd.org/base/head@137587 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f