Merge from, um, oblivion: vinumvar.h 1.49, vinumrequest.c 1.73
    When doing round-robin reads from a multi-plex volume, only switch to the
    next plex if the sector to be read isn't nearby the last read sector.

    Submitted by:  Vsevolod Lobko <seva@ip.net.ua> via ru@

Not only does this boost multi-access performance, but surprisingly it helps
single process sequential reads on my IDE disk mirror.  My guess is that the
read-a-block skip-a-block read-a-block effect of strict round-robin defeats
some sort of on-disk caching, leading to poor performance.  With this patch,
longer contiguous reads from a single disk are more common.

/sbin/vinum needs to be rebuilt as this patch changes the size of struct plex.

OK'd by: grog


git-svn-id: https://svn.freebsd.org/base/stable/4@138717 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f