Do more extensive checking of the userid field which is read in from the control-file for each print job.  This is partially because the previous checks still let through some characters which would cause trouble for other applications which try to process the resulting userid -- such as accounting programs.
But the main reason is to handle the case where some remote host sends a
print job where the given userid is an uppercase-version of the real userid.
For that case, lpd will now check for uppercase letters in the userid.  If
there are any, it will check to see if the given userid (with the uppercase
letters) is a valid one.  If it is *not* valid, then lpd will change the
userid to all-lowercase right when the job is received.

MFC after:	2 weeks


git-svn-id: https://svn.freebsd.org/base/head@139462 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f