Fix so all parts of lpd, lpc, lpq, and lprm will use the same algorithm for calculating the job number for a job based on the control-file name. We might receive cf-files named by other implementations of lpr, where the job number shown by lpq would not match the job number that other commands expected for the same name.
This also uses a newer algorithm for determining a job number, to avoid
problems caused when a control-file is named using an IP address, instead
of the hostname.

This also moved the declaration if isowner() from lp.h to rmjob.c.  When I
went to change the parameters, I noticed that rmjob.c was the only source
file which uses it.

MFC after:	2 weeks


git-svn-id: https://svn.freebsd.org/base/head@139464 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f