MFC src/sys/kern/kern_mbuf.c 1.6, src/sys/kern/uipc_mbuf.c 1.141, src/sys/sys/mbuf.h 1.165.  Original commit message:
"Optimize the way reference counting is performed with Mbufs.  We
do not need to perform an extra memory fetch in the Packet (Mbuf+Cluster)
constructor to initialize the reference counter anymore.  The reference
counts are located in a separate memory region (in the slab header,
because this zone is UMA_ZONE_REFCNT), so the memory fetch resulted very
often in a cache miss.  Additionally, and perhaps more significantly,
optimize the free mbuf+cluster (packet) case, which is very common, to
no longer require an atomic operation on free (to verify the reference
counter) if the reference on the cluster has never been increased (also
very common).  Reduces an atomic on mbuf free on average.

Original patch submitted by: Gerrit Nagelhout <gnagelhout@sandvine.com>"


git-svn-id: https://svn.freebsd.org/base/stable/5@141996 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f