The live CD itself no longer has /stand since a long time ago; however, there are still a couple of places under src/release such like fixit.profile assumed that system binary can be retrivied from /stand(ex: boot with the live CD and run "disklabel -e" in the fixit CD shell).
Since /stand is still functional in fixit floppy, and there are more than
one places in src/release needs to be updated(document for example) if we
want to make use of something like /mnt2/rescue/vi.  This commit try to
deal with aforementioned inconsistency with minimal effort by simply
create a symbolic link to /rescue.

Reviewed by:	jhb, sam (mentor)


git-svn-id: https://svn.freebsd.org/base/head@143999 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f