Add the '-S' and '-P' options.  The '-S' option can be used to split apart a string, and supports some text substitutions.  This can be used to provide all the flexibility (and more!) that was lost by recent changes to how the kernel parses #!-lines in shell scripts.
The '-P' option provides a way to specify an alternate set of directories
to use when searching for the 'utility' program to run.  This way you can
be sure what directories are used for that search, without changing the
value of PATH that the user has set.  Note that on FreeBSD 6.0, this
option is worthless unless the '-S' option is also used.

Approved by:	re (blanket `env')


git-svn-id: https://svn.freebsd.org/base/head@147493 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f