Add back missing copyright and license statement.  This is identical to the statement in ip_mroute.h, as well as being the same as what OpenBSD has done with this file.  It matches the copyright in NetBSD's 1.1 through 1.14 versions of the file as well, which they subsequently added back.
It appears to have been lost in the 4.4-lite1 import for FreeBSD 2.0,
but where and why I've not investigated further.  OpenBSD had the same
problem.  NetBSD had a copyright notice until Multicast 3.5 was
integrated verbatim back in 1995.  This appears to be the version that
made it into 4.4-lite1.

Approved by: re (scottl)
MFC after: 3 days


git-svn-id: https://svn.freebsd.org/base/head@147549 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f