Add fstab entries for ALTROOTSLICE and CFGSLICE, allowing us to do away with most of the hardcoding of device names in installed convenience scripts.
Provide a new script (root/updatep) to easily update the ALTROOTSLICE.  It
does the same as updatep1/updatep2 without hardcoded device names.  If noone
protests, I'd like to get rid of updatep1/updatep2 at some point.

As an extra bonus: turn on the error led while updating (on hardware with a
led).


git-svn-id: https://svn.freebsd.org/base/head@148377 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f