MFC:	etc/rc.d/jail	1.25
Skip jails which are already running and inform why.
We're checking for /var/run/jail_<name>.id file and if it exists, we don't
start the jail. It should be also safe in case of reboot(8), because
rc.d/cleanvar script is going to remove /var/run/jail_* files.

It helps to avoid potential mess when the same jail is started twice,
because of an administrator mistake (been there, done that).

Approved by:	re (kensmith)


git-svn-id: https://svn.freebsd.org/base/stable/6@149125 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f